[FNA] FNA Code Style: Tabbing, etc. - Media Dir
I *think* I've got a handle on the style now, and I've changed my MD formatting
prefs to be more in sync with your guidelines. 

Please tell me your thoughts on the MediaSource[] assignment
(MediaSource.cs, 44).  I split the assignment to multiple lines because of line
length.  I originally put the braces to separate lines (like you do with other
constructs), but with the new object being created inline it makes it a tiny bit
harder to grok that it is an array (IMHO).

If you'd prefer it be like:

MediaSource[] result = 
{ 
        new MediaSource("DumpMediaSource", MediaSourceType.LocalDevice)
};

then I will make that change.